% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uppy-dependencies.R
\name{uppy_dependencies}
\alias{uppy_dependencies}
\title{All Uppy dependencies}
\usage{
uppy_dependencies()
}
\value{
A list of \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} objects
}
\description{
Returns a list of all HTML dependencies required for Uppy.
}
\keyword{internal}
