% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uppy-dependencies.R
\name{uppy_dependency}
\alias{uppy_dependency}
\title{Uppy HTML dependencies}
\usage{
uppy_dependency()
}
\value{
An \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} object
}
\description{
Creates the HTML dependencies for Uppy.js. This function is called
automatically by \code{uppy_input()} and typically doesn't need to be
called directly.
}
\keyword{internal}
