% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uppy-server.R
\name{uppy_reset}
\alias{uppy_reset}
\title{Reset/clear an Uppy input}
\usage{
uppy_reset(input_id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{input_id}{Character. The input ID of the Uppy input to reset}

\item{session}{Shiny session object}
}
\value{
None (invisible NULL)
}
\description{
Programmatically reset an Uppy file input from the server, clearing
all selected files.
}
\examples{
\dontrun{
server <- function(input, output, session) {
  observeEvent(input$reset_button, {
    uppy_reset("files", session)
  })
}
}
}
