% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uppy-dependencies.R
\name{uppy_shiny_dependency}
\alias{uppy_shiny_dependency}
\title{Uppy Shiny binding dependencies}
\usage{
uppy_shiny_dependency()
}
\value{
An \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} object
}
\description{
Creates the HTML dependencies for the Uppy Shiny input binding.
This function is called automatically by \code{uppy_input()} and typically
doesn't need to be called directly.
}
\keyword{internal}
